Add files via upload
What was the problem you were solving in the projects for this course?
In this course we worked on two different projects. One was an online ordering system and platform for customers to be able to order from a store and the store can send the order for fulfillment from the supplier. We also had a project on the interaction between an ATM, Customer and Bank.
How did you approach the problem? Consider why data structures are important to understand.
For the online ordering system we had to gather what the store wanted to accomplish and had to create a multi platform system that had the ability to let customers order things, remove them from inventory, create lists, and update the supplier for shipment. We had to look at all the actors interactions to the website and start planning form there. 
For the ATM system we had to look at how each actor would interact with the different parts of the system and build it together in that order. There also had to be a higher level of security since this is peoples money, and access to personal information. 
How did you overcome any roadblocks you encountered while going through the activities or project?
When I hit a roadblock I was able to go back to previous weeks and reread through our resources and if I still couldn't get it I could watch YouTube videos that would explain more thoroughly why a piece of coding was written how it was and why it functioned the way it did. 
How has your work on this project expanded your approach to designing software and developing programs?
The work for these projects has definitely changed how I approach designing software since I had to really think of all the aspects but mainly the actors for the system. I'm used to thinking about the clients needs and building the code not necessarily every aspect of the system and who it interacts with. 
How has your work on this project evolved the way you write programs that are maintainable, readable, and adaptable?
My work on this project has evolved the way I write program since this class was more of a full stack class so I had to think of not only the building portion but also how it functioned for front end users which I'm not used to. so I had to work on full stack from beginning to end.